# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vantyx API
  version: 0.7.0
  description: |
    Vantyx バックエンドの公開 HTTP / WebSocket API 仕様。

    現時点でカバーしている機能:
    - ヘルスチェック (GET /healthz)
    - ローカルユーザー認証 (POST /api/login) と Cookie セッション
    - 自己情報取得 (GET /api/me)。応答に user_id, username, role を含む。
    - パスワード変更 (POST /api/me/password)。認証済みユーザーが自分のパスワードを変更。
    - UI ロケール設定 (PUT /api/me/locale)。`'' | en | ja` から選択し、サーバー側に永続化（ログイン応答 / GET /api/me の `locale` で取得）。
    - エラーレスポンスの本文 (`message`) は呼び出し元のロケールで返ります。解決順は (1) 認証済みユーザーの保存済みロケール、(2) `Accept-Language` ヘッダー、(3) デフォルト英語、です。
    - SSH 公開鍵（管理者のみ）: ユーザー管理でユーザーに紐づけて登録。GET/POST/DELETE /api/users/{user_id}/ssh-keys（一覧・追加・削除）。従来の /api/me/ssh-keys は自分用（管理者のみ）。Vantyx CLI（ssh user@vantyx）の公開鍵認証用。**これは接続先用の秘密鍵ライブラリ `/api/ssh-keys` とは別物です。**
    - 資格情報ライブラリ（管理者のみ、Keys / Identities 分離）: **Keys** — GET/POST/PUT/DELETE /api/ssh-keys（SSH 秘密鍵 PEM のみ、一覧は key_type / has_passphrase のみ返却）。**Identities** — GET/POST/PUT/DELETE /api/credential-identities（ユーザー名 + パスワード / 鍵 / 両方、auth_method 付き）。詳細は docs/credentials.md。
    - アクセスグループの一覧・作成 (GET/POST /api/groups)。一覧は limit / after_id でキーセットページネーション対応。
    - グループメンバー管理（管理者のみ）: GET/POST /api/groups/{group_id}/members、DELETE /api/groups/{group_id}/members/{user_id}。
    - タグ（グループ・ターゲット・ユーザー）: GET/PUT /api/groups/{group_id}/tags、/api/targets/{target_id}/tags、/api/users/{user_id}/tags。同じタグを持つユーザーはそのタグが付いたターゲット（またはそのタグが付いたグループに属するターゲット）にアクセス可能。
    - ユーザー登録・一覧（管理者のみ）: GET /api/users、POST /api/users。
...
servers:
...
  - url: /
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: vantyx_session
  schemas:
    ErrorResponse:
      type: object
      description: |
        エラー応答の汎用形式。`message` の文言は呼び出し元のロケール
        (`users.locale` → `Accept-Language` → 既定 `en`) で localize されて返ります。
        現在サポートしているロケールは `en` / `ja` です。
        全ての API ハンドラ（auth, users, groups, targets, sessions, recordings,
        files / file-transfers, tftp, settings など）の主要なエラー文言は
        i18n カタログ経由で生成されます。time_range クエリ (from/to)、
        protocol バリデーション、file-transfers の state/direction/backend/
        after_cursor バリデーション、`/api/me/password` のポリシーチェック、
        `proxyerrors`（TCP 接続失敗の SFTP / Telnet / SSH ターミナル
        メッセージ）、および `internal/access` / `internal/auth` の
        identifier / tag バリデーション (`validation.*`, `tags.*` キー群)
        も localize されます。残余のシステム / DB エラーは
        `common.internalError` で localize された 500 として返ります。
      properties:
        message:
          type: string
          description: 人間可読のエラーメッセージ（ロケーライズ済み）。
      required:
        - message
...
    Target:
      type: object
      description: 接続先ターゲット（サーバー）の情報。
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
          description: "階層管理用のフォルダパス（例: prod/network）。空の場合はルート。"
        host:
          type: string
          description: 接続先ホスト名または IP アドレス
        port:
          type: integer
          format: int32
          minimum: 1
          maximum: 65535
        protocol:
          type: string
          enum: ["ssh", "telnet", "vnc", "rdp", "ftp", "tftp"]
          description: "ssh=SSH ターミナル/SFTP, telnet=Telnet, vnc=VNC, rdp=RDP, ftp=FTP, tftp=TFTP"
        ssh_username:
          type: string
          description: 保存済み SSH ユーザー名（値は返す。パスワード・秘密鍵は含めない）
        has_stored_credentials:
          type: boolean
          description: SSH ユーザー名など資格情報が保存されている場合 true
        has_ssh_key:
          type: boolean
          description: SSH 秘密鍵が保存されている場合 true（値は含めない）
        needs_password:
          type: boolean
          description: ユーザー名のみ保存され接続時にパスワード入力が必要な場合 true
        needs_passphrase:
          type: boolean
          description: 暗号化された秘密鍵のみ保存され接続時にパスフレーズ入力が必要な場合 true
        sftp_enabled:
          type: boolean
        ftp_enabled:
          type: boolean
        tftp_enabled:
          type: boolean
        tags:
          type: array
          items:
            type: string
        ssh_host_key_fingerprint:
          type: string
          description: |
            登録済み SSH ホストキーの SHA-256 フィンガープリント。
            形式は `SHA256:<base64>`（OpenSSH 標準、`ssh-keygen -lf` と一致）。
            未登録の場合はフィールドが省略されるか空文字で返ります。
            初回接続時は WebSocket フレーム `host_key_unknown` が返り、
            UI 側が `PUT /api/targets/{id}/ssh-host-key` で採用できます。
          example: "SHA256:nThbg6kXUpJWGl7E1IGOCspRomTxdCARLviKw6E5SY8"
      required:
        - id
        - name
        - host
        - port
        - protocol
...
    AccessGroup:
      type: object
      description: アクセスグループ。ユーザーが所属し、ターゲットを紐づける単位。
      properties:
        id:
          type: string
          description: "階層を含むアクセスグループID（例: prod/network）"
        name:
          type: string
          description: 表示名
        targets:
          type: array
          items:
            $ref: '#/components/schemas/Target'
      required:
        - id
        - name
        - targets
    GroupListResponse:
      type: object
      description: limit または after_id 指定時のグループ一覧応答（キーセットページネーション）。
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccessGroup'
        next_cursor:
          type: string
          description: 次のページを取得する際に after_id に渡す値。存在しない場合は省略。
...
paths:
...
  /api/groups:
    get:
      summary: アクセスグループ一覧（所属 + ターゲット含む）
      description: |
        ログイン中のユーザーが所属するアクセスグループ一覧を返します。各グループには紐づくターゲット一覧が含まれます。
        **ページネーション**: クエリに limit または after_id を指定すると、レスポンスは `{ "items": [...], "next_cursor": "..." }` 形式になります。
        next_cursor が含まれる場合は、次リクエストで `after_id={next_cursor}` を付けて続きを取得します。limit 未指定時は 100、最大 1000。
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: 取得件数。指定時はキーセットページネーションになり、レスポンスは items/next_cursor 形式。
        - name: after_id
          in: query
          required: false
          schema:
            type: string
          description: この ID より大きいグループから取得（キーセット）。limit と組み合わせて使用。
      responses:
        '200':
          description: アクセスグループ一覧。limit/after_id 未指定時は配列、指定時は GroupListResponse。
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/AccessGroup'
                  - $ref: '#/components/schemas/GroupListResponse'
        '401':
          description: 未認証またはセッション無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: アクセスグループの追加
      description: 新しいアクセスグループを作成し、作成者をそのグループに追加します。
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                path:
                  type: string
                  description: "階層プレフィックス（例: prod）。省略可。"
              required:
                - name
      responses:
        '201':
          description: 作成されたアクセスグループ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessGroup'
        '400':
          description: リクエスト不正（name 不足、invalid request body など）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未認証またはセッション無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: グループ作成の内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...